refactor: convert http exception filter to all exceptions filter
